pyproject.toml
[project]
name = "djangotutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "crispy-bootstrap4>=2025.6",
    "django>=5.2.2",
    "django-crispy-forms>=2.4",
    "django-ratelimit>=4.1.0",
    "djangorestframework>=3.16.0",
    "httpie>=3.2.4",
    "pygments>=2.19.1",
]
